requirements.txt
# Core
pyyaml>=6.0
jinja2>=3.1
pydantic>=2.0
click>=8.0

# LLM APIs
openai>=1.0
anthropic>=0.30
tiktoken>=0.5

# Coverage processing & code parsing
tree-sitter>=0.21
tree-sitter-c>=0.21
tree-sitter-cpp>=0.22

# Random input generation (Phase 3)
exrex>=0.11
lxml>=4.9

# Statistical analysis (FuzzBench methodology)
pandas>=2.0
numpy>=1.24
scipy>=1.11
scikit-posthocs>=0.7

# Text similarity (contamination probe)
nltk>=3.8
Levenshtein>=0.21

# Plotting
matplotlib>=3.7
seaborn>=0.12

# Notebooks
jupyter>=1.0

# Testing
pytest>=7.0
pytest-timeout>=2.0

# Dev
ruff>=0.5

# Fine-tuning (Phase 4 only — install separately on GPU machine)
# torch>=2.1
# transformers>=4.36
# peft>=0.7
# bitsandbytes>=0.41
# datasets>=2.16
# accelerate>=0.25
